Add files via upload
A banking institution requires actionable insights into mortgage-backed securities, geographic business investment, and real estate analysis.  The mortgage bank would like to identify potential monthly mortgage expenses for each region based on monthly family income and rental of the real estate. A statistical model needs to be created to predict the potential demand in dollars amount of loan for each of the region in the USA.